fix(workspace): resolve duplicate module paths from file split (#2042)
PR #2040 split oversized files into module directories but left the
original .rs files alongside the new mod.rs files. Rust cannot have
both foo.rs and foo/mod.rs for the same module.

- Delete crates/symbolon/src/credential.rs (keep credential/mod.rs)
- Delete crates/organon/src/registry.rs (keep registry/mod.rs)
- Extract registry tests into registry/registry_tests.rs
- Rename original.rs → mod.rs for 6 theatron/tui split directories
- Add missing imports to test files (types lost via use super::*)
- Promote CachedFile visibility for cross-submodule test access
- Remove unused imports surfaced by -D warnings
- Fix formatting in renamed files

Closes #2042

Gate-Passed: kanon 0.1.0
