+++ crates/organon/src/types_tests.rs
@@ -3,6 +3,8 @@
     reason = "test: vec indices valid after asserting len"
 )]
 #![expect(clippy::expect_used, reason = "test assertions")]
+use std::collections::HashSet;
+
 use super::*;
 
 #[test]
