--- crates/symbolon/src/credential.rs
@@ -1,1000 +1,0 @@
-//! Credential provider implementations for LLM API key resolution.
-
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, RwLock};
-use std::time::{Duration, Instant, SystemTime};
-
-use serde::{Deserialize, Serialize};
-use tokio_util::sync::CancellationToken;
-use tracing::{Instrument, debug, error, info, warn};
-
-use aletheia_koina::credential::{Credential, CredentialProvider, CredentialSource};
-use aletheia_koina::secret::SecretString;
-
-use crate::circuit_breaker::{CircuitBreaker, CircuitBreakerConfig};
-use crate::encrypt::{
-    ENCRYPTED_SENTINEL, commit_key_file, decrypt, encrypt, load_or_create_key,
-    load_or_generate_key, prepare_key_file,
-};
-use crate::util::decode_jwt_exp_secs;
-
-/// Return current time as milliseconds since UNIX epoch, warning if the clock
-/// is before epoch rather than silently returning zero.
-fn unix_epoch_ms() -> u64 {
-    // WHY: as_millis() returns u128 but ms timestamps fit in u64 for ~500M years
-    let ms = SystemTime::now()
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .unwrap_or_else(|_| {
-            tracing::warn!("system clock before UNIX epoch, using epoch as fallback");
-            Duration::default()
-        })
-        .as_millis();
-    u64::try_from(ms).unwrap_or(u64::MAX)
-}
-
-/// Claude Code production OAuth client ID.
-const OAUTH_CLIENT_ID: &str = "9d1c250a-e61b-44d9-88ed-5944d1962f5e";
-
-/// OAuth token refresh endpoint.
-// WHY: must match console.anthropic.com, not platform.claude.com
-const OAUTH_TOKEN_URL: &str = "https://console.anthropic.com/v1/oauth/token";
-
-/// Refresh when token has less than this many seconds remaining.
-const REFRESH_THRESHOLD_SECS: u64 = 3600;
-
-/// Clock skew tolerance for token expiry checks (seconds).
-///
-/// WHY: clock differences between the OAuth provider and local system
-/// cause freshly obtained tokens to appear expired. 30 seconds is
-/// conservative enough to catch genuine expiry while tolerating
-/// typical NTP drift.
-const CLOCK_SKEW_LEEWAY_SECS: u64 = 30;
-
-/// How often the background refresh task checks token expiry.
-const REFRESH_CHECK_INTERVAL_SECS: u64 = 60;
-
-/// How often to check file mtime for external changes.
-const FILE_MTIME_CHECK_INTERVAL: Duration = Duration::from_secs(30);
-
-/// Minimum `expires_in` accepted from OAuth responses (seconds).
-const MIN_EXPIRES_IN_SECS: u64 = 60;
-
-/// Maximum `expires_in` accepted from OAuth responses (seconds).
-const MAX_EXPIRES_IN_SECS: u64 = 86400;
-
-/// On-disk credential file format.
-///
-/// Accepts both `"token"` (native format) and `"accessToken"` (Claude Code OAuth
-/// output) for backward compatibility. Serialization always writes `"token"`.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CredentialFile {
-    /// Access token (API key or OAuth access token).
-    #[serde(alias = "accessToken")]
-    pub token: String,
-    /// OAuth refresh token (absent for static API keys).
-    #[serde(rename = "refreshToken", skip_serializing_if = "Option::is_none")]
-    pub refresh_token: Option<String>,
-    /// Token expiry as milliseconds since epoch.
-    #[serde(rename = "expiresAt", skip_serializing_if = "Option::is_none")]
-    pub expires_at: Option<u64>,
-    /// OAuth scopes.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub scopes: Option<Vec<String>>,
-    /// Subscription tier.
-    #[serde(rename = "subscriptionType", skip_serializing_if = "Option::is_none")]
-    pub subscription_type: Option<String>,
-}
-
-impl CredentialFile {
-    /// Read and parse a credential file.
-    ///
-    /// Accepts three on-disk layouts:
-    ///
-    /// * **Encrypted**: prefixed with `ALETHEIA_ENC_V1:` -- decrypted using the
-    ///   sidecar key file (`<path>.key`) before parsing.
-    /// * **Flat**: `{"token": "...", "refreshToken": "..."}` (native) or with the
-    ///   `"accessToken"` alias produced by older Claude Code versions.
-    /// * **Wrapped**: `{"claudeAiOauth": {"accessToken": "...", ...}}`: the nested
-    ///   format written by current Claude Code releases.
-    ///
-    /// WHY: Claude Code changed its `.credentials.json` layout to nest all OAuth fields
-    /// under a `claudeAiOauth` top-level key. Without unwrapping it, fresh credentials
-    /// are invisible and the chain falls back to a stale env-var token.
-    #[must_use]
-    pub fn load(path: &Path) -> Option<Self> {
-        // WHY: orphaned .json.tmp files from crashed writes waste disk and confuse operators
-        let tmp = path.with_extension("json.tmp");
-        if tmp.exists() {
-            if let Err(e) = std::fs::remove_file(&tmp) {
-                warn!(error = %e, path = %tmp.display(), "failed to clean up orphaned temp file");
-            } else {
-                debug!(path = %tmp.display(), "cleaned up orphaned temp file");
-            }
-        }
-
-        let contents = std::fs::read_to_string(path).ok()?;
-
-        let json = if let Some(encoded) = contents.strip_prefix(ENCRYPTED_SENTINEL) {
-            // WHY: encrypted files must be decrypted before JSON parsing.
-            let key = load_or_create_key(path)
-                .map_err(|e| tracing::warn!(error = %e, path = %path.display(), "failed to load encryption key"))
-                .ok()?;
-            let plaintext = decrypt(&key, encoded.trim_end())
-                .map_err(|e| tracing::warn!(error = %e, path = %path.display(), "failed to decrypt credential file"))
-                .ok()?;
-            String::from_utf8(plaintext)
-                .map_err(|e| tracing::warn!(error = %e, "decrypted credential is not valid UTF-8"))
-                .ok()?
-        } else {
-            contents
-        };
-
-        if let Ok(cred) = serde_json::from_str::<Self>(&json) {
-            return Some(cred);
-        }
-
-        let outer: serde_json::Value = serde_json::from_str(&json).ok()?;
-        serde_json::from_value(outer.get("claudeAiOauth")?.clone()).ok()
-    }
-
-    /// Write the credential file atomically (write to temp, fsync, rename).
-    ///
-    /// The file is always encrypted with AES-256-GCM using a per-file key
-    /// stored in a sidecar `.key` file (mode 0600).
-    ///
-    /// Both the key file and credential file are written to temp files, fsynced,
-    /// and renamed atomically as a pair. An advisory write lock (`flock`) is held
-    /// for the duration to prevent races with Claude Code.
-    pub(crate) fn save(&self, path: &Path) -> std::io::Result<()> {
-        use std::io::Write as _;
-
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-
-        let json = serde_json::to_string_pretty(self).map_err(std::io::Error::other)?;
-
-        // WHY: load_or_generate_key tells us whether the key needs persisting so we
-        // can write both temp files before committing either rename
-        let (key, key_needs_persist) = load_or_generate_key(path)?;
-        let encoded = encrypt(&key, json.as_bytes())?;
-
-        let _lock = CredentialFileLock::exclusive(path)?;
-
-        // Phase 1: write all temp files and fsync
-        let key_tmp = if key_needs_persist {
-            Some(prepare_key_file(path, &key)?)
-        } else {
-            None
-        };
-
-        let cred_tmp = path.with_extension("json.tmp");
-        let mut file = std::fs::File::create(&cred_tmp)?;
-        file.write_all(ENCRYPTED_SENTINEL.as_bytes())?;
-        file.write_all(encoded.as_bytes())?;
-        file.flush()?;
-        file.sync_all()?;
-
-        // Phase 2: rename both atomically. If either fails, clean up both.
-        if let Some(ref ktmp) = key_tmp
-            && let Err(e) = commit_key_file(path, ktmp)
-        {
-            let _ = std::fs::remove_file(ktmp);
-            let _ = std::fs::remove_file(&cred_tmp);
-            return Err(e);
-        }
-
-        if let Err(e) = std::fs::rename(&cred_tmp, path) {
-            let _ = std::fs::remove_file(&cred_tmp);
-            // NOTE: key file was already renamed; it's valid on its own
-            return Err(e);
-        }
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(path, std::fs::Permissions::from_mode(0o600))?;
-        }
-        Ok(())
-    }
-
-    /// Whether this credential has a refresh token (OAuth flow).
-    #[must_use]
-    pub fn has_refresh_token(&self) -> bool {
-        self.refresh_token.as_ref().is_some_and(|t| !t.is_empty())
-    }
-
-    /// Seconds remaining until token expires. Returns `None` if no expiry set.
-    #[must_use]
-    pub fn seconds_remaining(&self) -> Option<i64> {
-        let expires_at_ms = self.expires_at?;
-        let now_ms = unix_epoch_ms();
-        // WHY: ms timestamps fit in i64 until year 292M; subtraction gives signed delta
-        let expires_i64 = i64::try_from(expires_at_ms).unwrap_or(i64::MAX);
-        let now_i64 = i64::try_from(now_ms).unwrap_or(i64::MAX);
-        Some((expires_i64 - now_i64) / 1000)
-    }
-
-    /// Whether the token needs refresh (expired or within threshold).
-    #[must_use]
-    #[expect(
-        dead_code,
-        reason = "refresh logic inlined in refresh_loop; kept as public API"
-    )]
-    pub(crate) fn needs_refresh(&self) -> bool {
-        match self.seconds_remaining() {
-            // WHY: REFRESH_THRESHOLD_SECS is a small constant that fits in i64
-            Some(remaining) => {
-                remaining < i64::try_from(REFRESH_THRESHOLD_SECS).unwrap_or(i64::MAX)
-            }
-            None => false,
-        }
-    }
-}
-
-/// Advisory file lock for credential read-modify-write cycles.
-///
-/// Uses `flock()` via `rustix` on a `.lock` sidecar file. The lock file is
-/// created alongside the credential file and never deleted (harmless).
-struct CredentialFileLock {
-    _file: std::fs::File,
-}
-
-impl CredentialFileLock {
-    /// Acquire a shared (read) lock on the credential file.
-    #[expect(
-        dead_code,
-        reason = "available for load() callers that need consistency"
-    )]
-    fn shared(credential_path: &Path) -> std::io::Result<Self> {
-        Self::lock(credential_path, rustix::fs::FlockOperation::LockShared)
-    }
-
-    /// Acquire an exclusive (write) lock on the credential file.
-    fn exclusive(credential_path: &Path) -> std::io::Result<Self> {
-        Self::lock(credential_path, rustix::fs::FlockOperation::LockExclusive)
-    }
-
-    #[cfg(unix)]
-    #[expect(
-        unsafe_code,
-        reason = "BorrowedFd::borrow_raw requires unsafe; fd is valid for File's lifetime"
-    )]
-    fn lock(credential_path: &Path, op: rustix::fs::FlockOperation) -> std::io::Result<Self> {
-        use std::os::unix::io::AsRawFd;
-
-        let lock_path = credential_path.with_extension("json.lock");
-        let file = std::fs::OpenOptions::new()
-            .create(true)
-            .truncate(false)
-            .write(true)
-            .open(&lock_path)?;
-        // WHY: flock is advisory but sufficient when all writers cooperate
-        rustix::fs::flock(
-            unsafe { rustix::fd::BorrowedFd::borrow_raw(file.as_raw_fd()) },
-            op,
-        )
-        .map_err(|e| std::io::Error::from_raw_os_error(e.raw_os_error()))?;
-        Ok(Self { _file: file })
-    }
-
-    #[cfg(not(unix))]
-    fn lock(credential_path: &Path, _op: rustix::fs::FlockOperation) -> std::io::Result<Self> {
-        let lock_path = credential_path.with_extension("json.lock");
-        let file = std::fs::OpenOptions::new()
-            .create(true)
-            .truncate(false)
-            .write(true)
-            .open(&lock_path)?;
-        Ok(Self { _file: file })
-    }
-}
-
-// NOTE: lock is released when _file is dropped (flock semantics)
-
-#[derive(Deserialize)]
-struct OAuthResponse {
-    access_token: String,
-    refresh_token: String,
-    #[serde(default = "default_expires_in")]
-    expires_in: u64,
-    scope: Option<String>,
-}
-
-impl std::fmt::Debug for OAuthResponse {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("OAuthResponse")
-            .field("access_token", &"[REDACTED]")
-            .field("refresh_token", &"[REDACTED]")
-            .field("expires_in", &self.expires_in)
-            .field("scope", &self.scope)
-            .finish()
-    }
-}
-
-fn default_expires_in() -> u64 {
-    28800 // NOTE: 8 hours
-}
-
-/// Clamp `expires_in` to [`MIN_EXPIRES_IN_SECS`, `MAX_EXPIRES_IN_SECS`].
-///
-/// WHY: a zero or negative value from a buggy OAuth server causes infinite
-/// refresh loops; an absurdly large value delays legitimate re-auth.
-fn clamp_expires_in(raw: u64) -> u64 {
-    let clamped = raw.clamp(MIN_EXPIRES_IN_SECS, MAX_EXPIRES_IN_SECS);
-    if clamped != raw {
-        warn!(
-            raw_expires_in = raw,
-            clamped_expires_in = clamped,
-            "OAuth expires_in outside [{MIN_EXPIRES_IN_SECS}, {MAX_EXPIRES_IN_SECS}], clamped"
-        );
-    }
-    clamped
-}
-
-/// OAuth token prefix used by Claude Code for OAuth access tokens.
-const OAUTH_TOKEN_PREFIX: &str = "sk-ant-oat";
-
-/// Reads a credential from an environment variable.
-///
-/// Automatically detects OAuth tokens by the `sk-ant-oat` prefix and
-/// returns [`CredentialSource::OAuth`] so callers use `Bearer` auth.
-pub struct EnvCredentialProvider {
-    var_name: String,
-    /// Force the credential source (e.g. OAuth for `ANTHROPIC_AUTH_TOKEN`).
-    force_source: Option<CredentialSource>,
-}
-
-impl EnvCredentialProvider {
-    /// Create a provider that reads credentials from the given environment variable.
-    #[must_use]
-    pub fn new(var_name: impl Into<String>) -> Self {
-        Self {
-            var_name: var_name.into(),
-            force_source: None,
-        }
-    }
-
-    /// Create a provider that always returns the given source type.
-    #[must_use]
-    pub fn with_source(var_name: impl Into<String>, source: CredentialSource) -> Self {
-        Self {
-            var_name: var_name.into(),
-            force_source: Some(source),
-        }
-    }
-}
-
-impl CredentialProvider for EnvCredentialProvider {
-    fn get_credential(&self) -> Option<Credential> {
-        std::env::var(&self.var_name).ok().and_then(|v| {
-            if v.is_empty() {
-                return None;
-            }
-
-            // WHY: static env var tokens cannot be refreshed; a refreshable file
-            // provider downstream must get a chance to supply a valid credential.
-            if v.starts_with(OAUTH_TOKEN_PREFIX)
-                && let Some(exp_secs) = decode_jwt_exp_secs(&v)
-            {
-                let now_secs = unix_epoch_ms() / 1000;
-                if exp_secs + CLOCK_SKEW_LEEWAY_SECS < now_secs {
-                    warn!(
-                        var = %self.var_name,
-                        exp_secs,
-                        now_secs,
-                        leeway_secs = CLOCK_SKEW_LEEWAY_SECS,
-                        "OAuth token from environment variable expired \
-                         (exp + leeway < now), falling through to next provider"
-                    );
-                    return None;
-                }
-            }
-
-            let source = self.force_source.clone().unwrap_or_else(|| {
-                if v.starts_with(OAUTH_TOKEN_PREFIX) {
-                    CredentialSource::OAuth
-                } else {
-                    CredentialSource::Environment
-                }
-            });
-            Some(Credential {
-                secret: SecretString::from(v),
-                source,
-            })
-        })
-    }
-
-    fn name(&self) -> &str {
-        &self.var_name
-    }
-}
-
-struct CachedFile {
-    token: SecretString,
-    mtime: SystemTime,
-    checked_at: Instant,
-}
-
-/// Reads a credential from a JSON file on disk.
-pub struct FileCredentialProvider {
-    path: PathBuf,
-    cached: RwLock<Option<CachedFile>>,
-}
-
-impl FileCredentialProvider {
-    /// Create a provider that reads credentials from the given JSON file path.
-    #[must_use]
-    pub fn new(path: PathBuf) -> Self {
-        Self {
-            path,
-            cached: RwLock::new(None),
-        }
-    }
-
-    fn current_mtime(&self) -> Option<SystemTime> {
-        std::fs::metadata(&self.path).ok()?.modified().ok()
-    }
-
-    fn reload(&self) -> Option<SecretString> {
-        let cred = CredentialFile::load(&self.path)?;
-        let mtime = self.current_mtime().unwrap_or_else(|| {
-            tracing::debug!(path = %self.path.display(), "could not read file mtime, using epoch");
-            SystemTime::UNIX_EPOCH
-        });
-
-        let token = SecretString::from(cred.token);
-        if let Ok(mut guard) = self.cached.write() {
-            *guard = Some(CachedFile {
-                token: token.clone(),
-                mtime,
-                checked_at: Instant::now(),
-            });
-        }
-        Some(token)
-    }
-}
-
-impl CredentialProvider for FileCredentialProvider {
-    fn get_credential(&self) -> Option<Credential> {
-        if let Ok(guard) = self.cached.read()
-            && let Some(cached) = guard.as_ref()
-        {
-            if cached.checked_at.elapsed() < FILE_MTIME_CHECK_INTERVAL {
-                return Some(Credential {
-                    secret: cached.token.clone(),
-                    source: CredentialSource::File,
-                });
-            }
-            if let Some(mtime) = self.current_mtime()
-                && mtime == cached.mtime
-            {
-                drop(guard);
-                if let Ok(mut w) = self.cached.write()
-                    && let Some(ref mut c) = *w
-                {
-                    c.checked_at = Instant::now();
-                }
-                if let Ok(g) = self.cached.read()
-                    && let Some(c) = g.as_ref()
-                {
-                    return Some(Credential {
-                        secret: c.token.clone(),
-                        source: CredentialSource::File,
-                    });
-                }
-            }
-        }
-
-        self.reload().map(|secret| Credential {
-            secret,
-            source: CredentialSource::File,
-        })
-    }
-
-    #[expect(
-        clippy::unnecessary_literal_bound,
-        reason = "trait requires &str return"
-    )]
-    fn name(&self) -> &str {
-        "file"
-    }
-}
-
-struct RefreshState {
-    current_token: SecretString,
-    refresh_token: SecretString,
-    expires_at_ms: u64,
-    subscription_type: Option<String>,
-}
-
-/// Wraps a credential file with background OAuth token refresh.
-///
-/// Cleanup is registered at construction time via [`CleanupRegistry`](aletheia_koina::cleanup::CleanupRegistry): the
-/// background task is cancelled and aborted when the provider is dropped,
-/// regardless of whether the drop occurs during normal execution, early
-/// error return, or panic unwind.
-pub struct RefreshingCredentialProvider {
-    /// Current OAuth token and refresh metadata. `None` after a fatal
-    /// refresh failure. Writers: the background refresh task (exclusive).
-    /// Readers: `get_credential()` on any thread. The `RwLock` ensures
-    /// readers never see a partially-updated token/expiry pair.
-    state: Arc<RwLock<Option<RefreshState>>>,
-    file_provider: FileCredentialProvider,
-    shutdown: CancellationToken,
-    /// Cleanup registered at task spawn time; fires on drop (LIFO order).
-    _cleanup: aletheia_koina::cleanup::CleanupRegistry,
-}
-
-impl RefreshingCredentialProvider {
-    /// Create a refreshing provider from a credential file path.
-    ///
-    /// Reads the credential file immediately and spawns a background refresh
-    /// task with a default circuit breaker. Requires a tokio runtime to be active.
-    #[must_use]
-    pub fn new(path: PathBuf) -> Option<Self> {
-        Self::with_circuit_breaker(path, CircuitBreakerConfig::default())
-    }
-
-    /// Create a refreshing provider with a custom circuit breaker configuration.
-    ///
-    /// Reads the credential file immediately and spawns a background refresh
-    /// task. Requires a tokio runtime to be active.
-    pub fn with_circuit_breaker(path: PathBuf, cb_config: CircuitBreakerConfig) -> Option<Self> {
-        let cred = CredentialFile::load(&path)?;
-        let refresh_token = cred.refresh_token.clone().filter(|t| !t.is_empty())?;
-
-        let state = Arc::new(RwLock::new(Some(RefreshState {
-            current_token: SecretString::from(cred.token.clone()),
-            refresh_token: SecretString::from(refresh_token),
-            expires_at_ms: cred.expires_at.unwrap_or_else(|| {
-                warn!(
-                    "credential has no expiry, treating as immediately expired to trigger refresh"
-                );
-                0
-            }),
-            subscription_type: cred.subscription_type,
-        })));
-
-        let shutdown = CancellationToken::new();
-        let circuit_breaker = Arc::new(CircuitBreaker::new(cb_config));
-
-        let task_state = Arc::clone(&state);
-        let task_shutdown = shutdown.clone();
-        let task_path = path.clone();
-        let task_cb = Arc::clone(&circuit_breaker);
-
-        let task = tokio::spawn(
-            async move {
-                refresh_loop(task_state, task_shutdown, task_path, task_cb).await;
-            }
-            .instrument(tracing::info_span!("credential_refresh")),
-        );
-
-        // Register cleanup at spawn time so the task is cancelled+aborted on
-        // drop even if construction is only partially completed in the caller.
-        let mut cleanup = aletheia_koina::cleanup::CleanupRegistry::new();
-        let shutdown_for_cleanup = shutdown.clone();
-        let abort_handle = task.abort_handle();
-        cleanup.register(move || {
-            shutdown_for_cleanup.cancel();
-            abort_handle.abort();
-        });
-
-        Some(Self {
-            state,
-            file_provider: FileCredentialProvider::new(path),
-            shutdown,
-            _cleanup: cleanup,
-        })
-    }
-
-    /// Signal the background refresh task to stop.
-    #[cfg_attr(
-        not(test),
-        expect(
-            dead_code,
-            reason = "called from tests; will be wired from server shutdown path"
-        )
-    )]
-    pub(crate) fn shutdown(&self) {
-        self.shutdown.cancel();
-    }
-}
-
-impl CredentialProvider for RefreshingCredentialProvider {
-    fn get_credential(&self) -> Option<Credential> {
-        if let Ok(guard) = self.state.read()
-            && let Some(ref s) = *guard
-            && !s.current_token.expose_secret().is_empty()
-        {
-            return Some(Credential {
-                secret: s.current_token.clone(),
-                source: CredentialSource::OAuth,
-            });
-        }
-        self.file_provider.get_credential()
-    }
-
-    #[expect(
-        clippy::unnecessary_literal_bound,
-        reason = "trait requires &str return"
-    )]
-    fn name(&self) -> &str {
-        "oauth"
-    }
-}
-
-// NOTE: No Drop impl -- cleanup is registered at setup time via CleanupRegistry.
-// The registry fires its callbacks (cancel token + abort task) on drop.
-
-/// Track last-observed mtime for file change detection.
-struct FileMtimeTracker {
-    last_mtime: Option<SystemTime>,
-}
-
-impl FileMtimeTracker {
-    fn new(path: &Path) -> Self {
-        let mtime = std::fs::metadata(path).ok().and_then(|m| m.modified().ok());
-        Self { last_mtime: mtime }
-    }
-
-    /// Returns `true` if the file's mtime has changed since last check.
-    fn has_changed(&mut self, path: &Path) -> bool {
-        let current = std::fs::metadata(path).ok().and_then(|m| m.modified().ok());
-        if current == self.last_mtime {
-            return false;
-        }
-        self.last_mtime = current;
-        true
-    }
-}
-
-/// Reload credentials from file when circuit is open and file has changed.
-fn try_reload_from_file(
-    state: &RwLock<Option<RefreshState>>,
-    path: &Path,
-    circuit_breaker: &CircuitBreaker,
-) {
-    let Some(file_cred) = CredentialFile::load(path) else {
-        return;
-    };
-    info!("credential file changed externally while circuit open, reloading");
-    if let Ok(mut guard) = state.write() {
-        *guard = Some(RefreshState {
-            current_token: SecretString::from(file_cred.token),
-            refresh_token: file_cred.refresh_token.map_or_else(
-                || {
-                    guard
-                        .as_ref()
-                        .map_or_else(|| SecretString::from(""), |s| s.refresh_token.clone())
-                },
-                SecretString::from,
-            ),
-            expires_at_ms: file_cred.expires_at.unwrap_or(0),
-            subscription_type: file_cred.subscription_type,
-        });
-    }
-    circuit_breaker.reset();
-}
-
-/// Build the post-refresh state, adopting the on-disk credential if newer.
-fn resolve_post_refresh_state(
-    path: &Path,
-    resp: OAuthResponse,
-    new_expires_at_ms: u64,
-    subscription_type: Option<String>,
-) -> RefreshState {
-    if let Some(on_disk) = CredentialFile::load(path)
-        && on_disk.expires_at.unwrap_or(0) > new_expires_at_ms
-    {
-        info!(
-            our_expiry = new_expires_at_ms,
-            file_expiry = on_disk.expires_at.unwrap_or(0),
-            "file has newer credential than our refresh, adopting"
-        );
-        RefreshState {
-            current_token: SecretString::from(on_disk.token),
-            refresh_token: on_disk.refresh_token.map_or_else(
-                || SecretString::from(resp.refresh_token),
-                SecretString::from,
-            ),
-            expires_at_ms: on_disk.expires_at.unwrap_or(new_expires_at_ms),
-            subscription_type: on_disk.subscription_type,
-        }
-    } else {
-        RefreshState {
-            current_token: SecretString::from(resp.access_token),
-            refresh_token: SecretString::from(resp.refresh_token),
-            expires_at_ms: new_expires_at_ms,
-            subscription_type,
-        }
-    }
-}
-
-async fn refresh_loop(
-    state: Arc<RwLock<Option<RefreshState>>>,
-    shutdown: CancellationToken,
-    path: PathBuf,
-    circuit_breaker: Arc<CircuitBreaker>,
-) {
-    let client = reqwest::Client::new();
-    let check_interval = Duration::from_secs(REFRESH_CHECK_INTERVAL_SECS);
-    let mut mtime_tracker = FileMtimeTracker::new(&path);
-
-    loop {
-        tokio::select! {
-            biased;
-            () = shutdown.cancelled() => {
-                info!("credential refresh loop shutting down");
-                break;
-            }
-            () = tokio::time::sleep(check_interval) => {}
-        }
-
-        // FIX 4: when circuit is open, poll file for external credential updates
-        if !circuit_breaker.check_allowed() {
-            if mtime_tracker.has_changed(&path) {
-                try_reload_from_file(&state, &path, &circuit_breaker);
-            } else {
-                debug!(
-                    state = %circuit_breaker.state(),
-                    "OAuth refresh circuit breaker is open, skipping refresh attempt"
-                );
-            }
-            continue;
-        }
-
-        let (refresh_token_value, subscription_type, expires_at_ms, needs_refresh) = {
-            let Ok(guard) = state.read() else {
-                continue;
-            };
-            let Some(s) = guard.as_ref() else {
-                continue;
-            };
-            let now_ms = unix_epoch_ms();
-            let expires_i64 = i64::try_from(s.expires_at_ms).unwrap_or(i64::MAX);
-            let now_i64 = i64::try_from(now_ms).unwrap_or(i64::MAX);
-            let remaining_secs = (expires_i64 - now_i64) / 1000;
-            let threshold = i64::try_from(REFRESH_THRESHOLD_SECS).unwrap_or(i64::MAX);
-            (
-                s.refresh_token.expose_secret().to_owned(),
-                s.subscription_type.clone(),
-                s.expires_at_ms,
-                remaining_secs < threshold,
-            )
-        };
-
-        if !needs_refresh {
-            continue;
-        }
-
-        info!(
-            expires_at_ms,
-            now_ms = unix_epoch_ms(),
-            "credential refresh needed"
-        );
-
-        if let Some(resp) = do_refresh(&client, &refresh_token_value).await {
-            circuit_breaker.record_success();
-            let expires_in = clamp_expires_in(resp.expires_in);
-            let new_expires_at_ms = unix_epoch_ms() + expires_in * 1000;
-
-            let scopes = resp
-                .scope
-                .as_deref()
-                .map(|s| s.split_whitespace().map(String::from).collect());
-            let cred_file = CredentialFile {
-                token: resp.access_token.clone(),
-                refresh_token: Some(resp.refresh_token.clone()),
-                expires_at: Some(new_expires_at_ms),
-                scopes,
-                subscription_type: subscription_type.clone(),
-            };
-
-            match cred_file.save(&path) {
-                Ok(()) => {
-                    mtime_tracker.has_changed(&path);
-                    let final_state = resolve_post_refresh_state(
-                        &path,
-                        resp,
-                        new_expires_at_ms,
-                        subscription_type,
-                    );
-                    if let Ok(mut guard) = state.write() {
-                        *guard = Some(final_state);
-                    }
-                    crate::metrics::record_token_refresh(true);
-                    info!(expires_in_secs = expires_in, "OAuth token refreshed");
-                }
-                Err(e) => {
-                    error!(error = %e, "failed to write refreshed credential file, keeping previous in-memory token");
-                    crate::metrics::record_credential_write_failure();
-                    crate::metrics::record_token_refresh(true);
-                }
-            }
-        } else {
-            circuit_breaker.record_failure();
-            crate::metrics::record_token_refresh(false);
-            warn!("OAuth token refresh failed, will retry next cycle");
-        }
-    }
-}
-
-async fn do_refresh(client: &reqwest::Client, refresh_token: &str) -> Option<OAuthResponse> {
-    // WHY: Anthropic OAuth endpoint expects form-urlencoded, not JSON
-    let body = format!(
-        "grant_type=refresh_token&refresh_token={refresh_token}&client_id={OAUTH_CLIENT_ID}",
-    );
-
-    let resp = client
-        .post(OAUTH_TOKEN_URL)
-        .header("Content-Type", "application/x-www-form-urlencoded")
-        .body(body)
-        .timeout(Duration::from_secs(30))
-        .send()
-        .await
-        .map_err(|e| {
-            warn!(error = %e, "OAuth refresh request failed");
-            e
-        })
-        .ok()?;
-
-    if !resp.status().is_success() {
-        let status = resp.status();
-        let body = resp.text().await.unwrap_or_else(|e| {
-            warn!("failed to read OAuth error response body: {e}");
-            String::new()
-        });
-        warn!(status = %status, body = %body, "OAuth refresh returned error");
-        return None;
-    }
-
-    resp.json::<OAuthResponse>()
-        .await
-        .map_err(|e| {
-            warn!(error = %e, "failed to parse OAuth refresh response");
-            e
-        })
-        .ok()
-}
-
-/// Force a one-shot OAuth token refresh (for CLI `credential refresh`).
-///
-/// # Errors
-///
-/// Returns an error if the credential file cannot be read, contains no refresh
-/// token, the OAuth refresh request fails, or the updated credentials cannot
-/// be saved.
-#[must_use = "refreshed credentials must be used or persisted"]
-pub async fn force_refresh(path: &Path) -> Result<CredentialFile, String> {
-    let cred = CredentialFile::load(path)
-        .ok_or_else(|| format!("cannot read credential file: {}", path.display()))?;
-
-    let refresh_token = cred
-        .refresh_token
-        .as_ref()
-        .filter(|t| !t.is_empty())
-        .ok_or("no refresh token in credential file")?;
-
-    let client = reqwest::Client::new();
-    let resp = do_refresh(&client, refresh_token)
-        .await
-        .ok_or("OAuth refresh failed")?;
-
-    let expires_in = clamp_expires_in(resp.expires_in);
-    let now_ms = unix_epoch_ms();
-    let expires_at_ms = now_ms + expires_in * 1000;
-
-    let scopes = resp
-        .scope
-        .map(|s| s.split_whitespace().map(String::from).collect());
-    let updated = CredentialFile {
-        token: resp.access_token,
-        refresh_token: Some(resp.refresh_token),
-        expires_at: Some(expires_at_ms),
-        scopes,
-        subscription_type: cred.subscription_type,
-    };
-    updated
-        .save(path)
-        .map_err(|e| format!("failed to write credential file: {e}"))?;
-
-    Ok(updated)
-}
-
-/// Default path to the Claude Code credentials file.
-///
-/// Returns `~/.claude/.credentials.json`, resolving `~` via `$HOME`.
-/// Returns `None` if `$HOME` is not set.
-#[must_use]
-pub fn claude_code_default_path() -> Option<PathBuf> {
-    std::env::var_os("HOME").map(|home| {
-        PathBuf::from(home)
-            .join(".claude")
-            .join(".credentials.json")
-    })
-}
-
-/// Build a credential provider from a Claude Code credentials file.
-///
-/// If the file contains a refresh token, returns a [`RefreshingCredentialProvider`]
-/// that keeps the token fresh in the background. Otherwise returns a
-/// [`FileCredentialProvider`] for static token reads.
-///
-/// Returns `None` if the file does not exist or cannot be parsed.
-#[must_use]
-pub fn claude_code_provider(path: &Path) -> Option<Box<dyn CredentialProvider>> {
-    claude_code_provider_with_config(path, CircuitBreakerConfig::default())
-}
-
-/// Build a credential provider with a custom circuit breaker configuration.
-///
-/// See [`claude_code_provider`] for behavior details.
-pub fn claude_code_provider_with_config(
-    path: &Path,
-    cb_config: CircuitBreakerConfig,
-) -> Option<Box<dyn CredentialProvider>> {
-    if !path.exists() {
-        return None;
-    }
-    let cred = CredentialFile::load(path)?;
-    if cred.has_refresh_token()
-        && let Some(refreshing) =
-            RefreshingCredentialProvider::with_circuit_breaker(path.to_path_buf(), cb_config)
-    {
-        info!(
-            path = %path.display(),
-            "Claude Code credentials found (OAuth auto-refresh)"
-        );
-        return Some(Box::new(refreshing));
-    }
-    info!(
-        path = %path.display(),
-        "Claude Code credentials found (static token)"
-    );
-    Some(Box::new(FileCredentialProvider::new(path.to_path_buf())))
-}
-
-/// Ordered list of credential providers. First to return `Some` wins.
-pub struct CredentialChain {
-    providers: Vec<Box<dyn CredentialProvider>>,
-    resolved_name: RwLock<String>,
-}
-
-impl CredentialChain {
-    /// Create a credential chain that tries each provider in order until one succeeds.
-    #[must_use]
-    pub fn new(providers: Vec<Box<dyn CredentialProvider>>) -> Self {
-        Self {
-            providers,
-            resolved_name: RwLock::new("chain".to_owned()),
-        }
-    }
-}
-
-impl CredentialProvider for CredentialChain {
-    fn get_credential(&self) -> Option<Credential> {
-        for provider in &self.providers {
-            if let Some(cred) = provider.get_credential() {
-                if let Ok(mut name) = self.resolved_name.write() {
-                    provider.name().clone_into(&mut *name);
-                }
-                return Some(cred);
-            }
-        }
-        None
-    }
-
-    #[expect(
-        clippy::unnecessary_literal_bound,
-        reason = "trait requires &str return"
-    )]
-    fn name(&self) -> &str {
-        "chain"
-    }
-}
-
-#[cfg(test)]
-#[path = "credential_tests.rs"]
-mod tests;
