+++ crates/theatron/tui/src/state/memory/mod.rs
