+++ crates/theatron/tui/src/update/streaming/mod.rs
