+++ crates/theatron/tui/src/view/chat/mod.rs
