+++ crates/theatron/tui/src/view/memory/mod.rs
